feat(enum): implement Default trait for BatchMode, KeyPairType, and XcmVersion enums

+++ apex-sdk-substrate/src/lib.rs
@@ -317,7 +317,7 @@ impl SubstrateAdapter {
                         }
                     }
 
-                    let confirmations = (latest_number - block_num) as u32;
+                    let confirmations = latest_number - block_num;
 
                     return if success {
                         Ok(TransactionStatus::Confirmed {
